backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
anthropic==0.34.2
chromadb==0.5.11
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis[hiredis]==5.1.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pymupdf==1.24.11
python-docx==1.1.2
sentence-transformers==3.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.4.0
tenacity==9.0.0
tiktoken==0.7.0
langdetect==1.0.9
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
